pyproject.toml
[project]
name = "hd-google-hackathon"
version = "0.1.0"
description = "Agentic platform that unifies dealer support workflows with orchestrated AI agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
